blanchsb added popup for image extension color management and texture wrapping
